Fix continue only if action + add tests
