+++ server/test/lib/scene/actions/scene.action.continueOnlyIf.test.js
@@ -1,0 +1,124 @@
+const { fake } = require('sinon');
+const EventEmitter = require('events');
+const chaiAssert = require('chai').assert;
+
+const { ACTIONS } = require('../../../../utils/constants');
+const { AbortScene } = require('../../../../utils/coreErrors');
+const { executeActions } = require('../../../../lib/scene/scene.executeActions');
+
+const StateManager = require('../../../../lib/state');
+
+const event = new EventEmitter();
+
+describe('scene.continue-only-if', () => {
+  it('should abort scene, condition is not verified', async () => {
+    const stateManager = new StateManager(event);
+    stateManager.setState('deviceFeature', 'my-device-feature', {
+      category: 'light',
+      type: 'binary',
+      last_value: 15,
+    });
+    const device = {
+      setValue: fake.resolves(null),
+    };
+    const scope = {};
+    const promise = executeActions(
+      { stateManager, event, device },
+      [
+        [
+          {
+            type: ACTIONS.DEVICE.GET_VALUE,
+            device_feature: 'my-device-feature',
+          },
+        ],
+        [
+          {
+            type: ACTIONS.CONDITION.ONLY_CONTINUE_IF,
+            conditions: [
+              {
+                variable: '0.0.last_value',
+                operator: '=',
+                value: 20,
+              },
+            ],
+          },
+        ],
+      ],
+      scope,
+    );
+    return chaiAssert.isRejected(promise, AbortScene);
+  });
+  it('should finish scene, condition is verified', async () => {
+    const stateManager = new StateManager(event);
+    stateManager.setState('deviceFeature', 'my-device-feature', {
+      category: 'light',
+      type: 'binary',
+      last_value: 15,
+    });
+    const device = {
+      setValue: fake.resolves(null),
+    };
+    const scope = {};
+    await executeActions(
+      { stateManager, event, device },
+      [
+        [
+          {
+            type: ACTIONS.DEVICE.GET_VALUE,
+            device_feature: 'my-device-feature',
+          },
+        ],
+        [
+          {
+            type: ACTIONS.CONDITION.ONLY_CONTINUE_IF,
+            conditions: [
+              {
+                variable: '0.0.last_value',
+                operator: '=',
+                value: 15,
+              },
+            ],
+          },
+        ],
+      ],
+      scope,
+    );
+  });
+  it('should finish scene, condition is verified', async () => {
+    const stateManager = new StateManager(event);
+    const http = {
+      request: fake.resolves({ result: [18], error: null }),
+    };
+    const device = {
+      setValue: fake.resolves(null),
+    };
+    const scope = {};
+    await executeActions(
+      { stateManager, event, device, http },
+      [
+        [
+          {
+            type: ACTIONS.HTTP.REQUEST,
+            method: 'post',
+            url: 'http://test.test',
+            body: '{"toto":"toto"}',
+            headers: [],
+          },
+        ],
+        [
+          {
+            type: ACTIONS.CONDITION.ONLY_CONTINUE_IF,
+            conditions: [
+              {
+                variable: '0.0.result.[0]',
+                operator: '=',
+                value: 18,
+              },
+            ],
+          },
+        ],
+      ],
+      scope,
+    );
+  });
+});
